Move RewriteProjectParser to top-level package
