+++ spring-rewrite-commons-examples/list-applicable-recipes-example/src/main/java/org/example/app/ListApplicableRecipesExampleApplication.java
@@ -27,7 +27,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.core.io.Resource;
 import org.springframework.rewrite.parser.ProjectScanner;
-import org.springframework.rewrite.parser.RewriteProjectParser;
+import org.springframework.rewrite.RewriteProjectParser;
 import org.springframework.rewrite.parser.RewriteProjectParsingResult;
 import org.springframework.rewrite.recipes.RewriteRecipeDiscovery;
