+++ spring-rewrite-commons-launcher/src/main/java/org/springframework/rewrite/execution/RewriteRecipeLauncher.java
@@ -19,7 +19,7 @@ import org.jetbrains.annotations.NotNull;
 import org.openrewrite.Recipe;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.rewrite.parser.RewriteProjectParser;
+import org.springframework.rewrite.RewriteProjectParser;
 import org.springframework.rewrite.parser.RewriteProjectParsingResult;
 import org.springframework.rewrite.resource.ProjectResourceSet;
 import org.springframework.rewrite.resource.ProjectResourceSetFactory;
